fix: correctly handle indexing annotated tags
Before, we would see a lot of zero value timestamps for the
`CommittedDate` field for tags from multiple repos. This was because
`commitedDate` field in the github graphql schema is only used for
lightweight tags. Annotated tags store their creation date in a
different `Tagger.Date` field.

This commit adds support for indexing annotated tags by correctly
checking both of the fields and setting the tag timestamp to whichever
field is not zero.