fix: remove hardcoded references to the VCS hostname (#34)
Before, we were hardcoding references to the Netflix VCS hostname in the
github client and in the server. This commit makes a change to rely on
the existing githubHostName flag we're already requiring as part of
github client initialization and to use the string provided there
instead. This should allow us to support arbitrary GitHub enterprise
hostnames.

For now, we're also passing githubHostName to the server so that we can
prepend module paths with the hostname when serving index results. This
is an okay short-term solution but should be fixed properly by
extracting and storing module path from go.mod rather than just the VCS
repo path. This will be addressed in the follow up work.