feat: separate Go module paths from VCS repo paths
Before, our indexer was assuming that Go module paths always matched VCS
repo paths. That's usually a good assumption. However, if a Go module
started in one VCS and moved to a different VCS, or if the author of the
Go module hasn't followed typical conventions, that led to us returning
wrong module paths from the index.

This commit separates the concept of the Go module path from the VCS
repo path. To do that, we look up go.mod file from the repo root and
attempt to look up Go module path from that file. This doesn't work for
all of the use-cases but should cover the typical case of a Go repo. For
other setups, like Go monorepos that might have a lot of modules stored
in the repository, we fallback to the previous assumption that the repo
module matches the VCS path.