feat: skip tags with present but invalid go.mod paths
Before, we would default to github repo module path for modules with
invalid go.mod files. This was not great because these modules couldn't
be imported by clients regardless.

This commit adds an explicit check for the validity of module path. If
the module path is invalid, we skip storing that tag for the module
entirely.